Introduce model adapter and remove ad-hoc helpers
