+++ vllm_metal/v1/model_runner.py
@@ -70,10 +70,7 @@ from vllm_metal.v1.contiguous_cache import (
     _extract_kv_cache,
     _merge_kv_caches,
 )
-from vllm_metal.v1.model_compat import (
-    resolve_max_head_dim,
-    should_force_text_backbone,
-)
+from vllm_metal.v1.model_adapter import DefaultModelAdapter, ModelAdapter
 from vllm_metal.v1.sampling_batch import (
     GREEDY_TEMPERATURE_EPS,
     SamplingBatch,
@@ -81,7 +78,6 @@ from vllm_metal.v1.sampling_batch import (
     sample_from_logits,
     sample_prefill_tokens,
 )
-from vllm_metal.v1.vlm_utils import _vlm_text_model
 
 logger = init_logger(__name__)
 
@@ -221,6 +217,7 @@ class MetalModelRunner:
         self.use_async_scheduling = bool(self.scheduler_config.async_scheduling)
         self.device = device
         self.metal_config = get_config()
+        self._model_adapter: ModelAdapter = DefaultModelAdapter()
 
         self.model: Any = None
         self.tokenizer: Any = None
@@ -317,7 +314,7 @@ class MetalModelRunner:
         the upstream pattern, and is a separate future effort.
         """
         if self._is_vlm:
-            return _vlm_text_model(self.model)
+            return self._model_adapter.text_model(self.model)
         return self.model
 
     @property
@@ -368,7 +365,9 @@ class MetalModelRunner:
             True if the model is multimodal/VLM, False otherwise
         """
         hf_config = getattr(self.model_config, "hf_config", None)
-        if hf_config is not None and should_force_text_backbone(hf_config):
+        if hf_config is not None and self._model_adapter.should_force_text_backbone(
+            hf_config
+        ):
             return False
         if hasattr(self.model_config, "is_multimodal_model"):
             return self.model_config.is_multimodal_model
@@ -564,7 +563,7 @@ class MetalModelRunner:
             if hidden_size and num_attention_heads
             else None
         )
-        head_dim = resolve_max_head_dim(args, head_dim)
+        head_dim = self._model_adapter.resolve_max_head_dim(args, head_dim)
 
         # Fail fast if critical dims are missing
         missing = []
