+++ vllm_metal/v1/worker.py
@@ -32,7 +32,7 @@ from vllm_metal.paged_attention_backend.mla import MLAPagedAttentionBackend
 from vllm_metal.platform import MetalPlatform
 from vllm_metal.stt.policy import STT_SCHED_AVAILABLE_BYTES
 from vllm_metal.utils import set_wired_limit
-from vllm_metal.v1.model_compat import require_uniform_kv_heads
+from vllm_metal.v1.model_adapter import DefaultModelAdapter
 
 if TYPE_CHECKING:
     from vllm_metal.v1.model_runner import (
@@ -41,6 +41,7 @@ if TYPE_CHECKING:
     )
 
 logger = init_logger(__name__)
+_MODEL_ADAPTER = DefaultModelAdapter()
 
 
 def init_worker_distributed_environment(
@@ -177,7 +178,7 @@ class MetalWorker(WorkerBase):
         max_model_len.
         """
         runner = self.model_runner
-        require_uniform_kv_heads(runner.model_args, runner.num_kv_heads)
+        _MODEL_ADAPTER.require_uniform_kv_heads(runner.model_args, runner.num_kv_heads)
 
         # Use cache_config.block_size (not metal_config) because vLLM's
         # hybrid alignment may have adjusted it to match mamba page size.
